Video generation from hcs Images by timepoints - remarks fix + uuid for video folder
