+++ deploy/docker/cp-tools/research/cellprofiler-web-api/app/src/hcs_manager.py
@@ -5,17 +5,6 @@ from time import sleep
 import glob
 import os
 from .z_planes_pipeline import ZPlanesPipeline
-from tifffile import imread
-import numpy as np
-from moviepy.editor import ImageSequenceClip
-import time
-from PIL import Image, ImageOps
-import xml.etree.ElementTree as ET
-import json
-import errno
-
-CELLPROFILER_API_COMMON_RESULTS_DIR = os.getenv('CELLPROFILER_API_COMMON_RESULTS_DIR')
-HCS_CLOUD_FILES_SCHEMA = os.getenv('HCS_PARSING_CLOUD_FILES_SCHEMA', 's3')
 
 
 class HCSManager:
@@ -164,223 +153,16 @@ class HCSManager:
             parent_pipeline_inputs.append(parent_pipeline.map_to_file_name(image))
         return parent_pipeline_inputs
 
-    def create_clip(self, params):
-        t = time.time()
-
-        clip_format = params.get('format') if 'format' in params else '.webm'
-        codec = params.get('codec') if 'codec' in params else ('mpeg4' if clip_format == '.mp4' else None)
-        fps = int(params.get('fps')) if 'fps' in params else 1
-        duration = float(params.get('duration')) if 'duration' in params else 1
-        if duration < 1:
-            raise RuntimeError('Duration should be >= 1')
-        sequence_id = params.get('sequenceId') if 'sequenceId' in params else None
-        plane_id = params.get('planeId') if 'planeId' in params else '1'
-
-        path = self._get_required_field(params, 'path')
-        path = self.extract_local_path(path)
-        by_field = int(self._get_required_field(params, 'byField'))
-        cell = int(self._get_required_field(params, 'cell'))
-
-        preview_dir, sequences = self.parse_hcs(path, sequence_id)
-        preview_dir = self.extract_local_path(preview_dir)
-        index_path = preview_dir + '/index.xml'
-        all_channels = self.get_all_channels(index_path)
-        planes = self.get_planes(index_path)
-        if plane_id not in planes:
-            raise RuntimeError('Incorrect Z plane id [{}]'.format(plane_id))
-        channels_num = len(all_channels)
-        selected_channel_ids, selected_channels = self.get_selected_channels(params, all_channels)
-
-        clips = []
-        durations = []
-        for seq in sequences.keys():
-            timepoints = len(sequences[seq])
-            ome_tiff_path = self.get_ome_tiff_path(preview_dir, seq, by_field)
-            pages = self.get_pages(selected_channel_ids, plane_id, planes, channels_num, timepoints, cell)
-            images = self.read_images(ome_tiff_path, pages)
-            curr = 0
-            for time_point in range(timepoints):
-                channel_images = []
-                for channel in selected_channels:
-                    channel_image = images[curr]
-                    colored_image = self.color_image(channel_image, channel)
-                    channel_images.append(colored_image)
-                    curr = curr + 1
-                merged_image = self.merge_channels(channel_images)
-                clips.append(np.array(merged_image))
-                durations.append(duration)
-
-        clip_path = CELLPROFILER_API_COMMON_RESULTS_DIR + self.get_clip_file_dir(path)
-        self._mkdir(clip_path)
-        clip_name = self.get_clip_file_name(by_field, cell, clip_format, sequence_id, plane_id)
-        clip_full_path = clip_path + clip_name
-        slide = ImageSequenceClip(clips, durations=durations)
-        slide.write_videofile(clip_full_path, codec=codec, fps=fps)
-        t1 = time.time()
-        return clip_full_path, round((t1 - t), 2)
-
-    @staticmethod
-    def read_images(ome_tiff_path, pages):
-        images = imread(ome_tiff_path, key=pages)
-        if len(pages) == 1:
-            images = np.reshape(images, (1, 1080, 1080))
-        return images
-
-    @staticmethod
-    def get_clip_file_dir(path):
-        hcs_file_base_name = os.path.basename(path)
-        hcs_file_name = os.path.splitext(hcs_file_base_name)[0]
-        return '/video/{}/'.format(hcs_file_name)
-
-    @staticmethod
-    def get_clip_file_name(by_field, cell, clip_format, sequence_id, plane_id):
-        cell_prefix = ('field{}' if by_field else '/well{}').format(str(cell))
-        plane_prefix = 'plane{}'.format(plane_id)
-        sequence_prefix = ('' if sequence_id is None else 'seq{}').format(sequence_id)
-        return cell_prefix + sequence_prefix + plane_prefix + clip_format
-
-    @staticmethod
-    def get_pages(channel_ids, plane_id, planes, channels, timepoints, cell):
-        pages = []
-        num = 0
-        for time_point in range(timepoints):
-            for plane in planes:
-                for channel_id in range(channels):
-                    if (channel_id in channel_ids) and (plane == plane_id):
-                        pages.append(num + len(planes) * channels * timepoints * cell)
-                    num = num + 1
-        return pages
-
-    @staticmethod
-    def extract_xml_schema(xml_info_root):
-        full_schema = xml_info_root.tag
-        return full_schema[:full_schema.rindex('}') + 1]
-
-    @staticmethod
-    def parse_hcs(path, sequence_id):
-        with open(path) as json_file:
-            data = json.load(json_file)
-            preview_dir = data['previewDir']
-            time_series_details = data['time_series_details']
-            sequences = dict()
-            if sequence_id is not None:
-                sequences[sequence_id] = time_series_details[sequence_id]
-            else:
-                sequences = time_series_details
-            return preview_dir, sequences
-
-    def get_all_channels(self, path):
-        tree = ET.parse(path)
-        hcs_xml_info_root = tree.getroot()
-        hcs_schema_prefix = self.extract_xml_schema(hcs_xml_info_root)
-        maps = hcs_xml_info_root.find(hcs_schema_prefix + 'Maps').findall(hcs_schema_prefix + 'Map')
-        channels = []
-        for m in maps:
-            entries = m.findall(hcs_schema_prefix + 'Entry')
-            for e in entries:
-                channel_name = e.find(hcs_schema_prefix + 'ChannelName')
-                if channel_name is not None:
-                    channel = dict()
-                    channel['id'] = int(e.attrib.get('ChannelID')) - 1
-                    channel['name'] = channel_name.text
-                    channels.append(channel)
-        return channels
-
-    def get_planes(self, path):
-        tree = ET.parse(path)
-        hcs_xml_info_root = tree.getroot()
-        hcs_schema_prefix = self.extract_xml_schema(hcs_xml_info_root)
-        images = hcs_xml_info_root.find(hcs_schema_prefix + 'Images').findall(hcs_schema_prefix + 'Image')
-        planes = []
-        for i in images:
-            plane_id = i.find(hcs_schema_prefix + 'PlaneID').text
-            if plane_id not in planes:
-                planes.append(plane_id)
-        return planes
-
-    @staticmethod
-    def extract_local_path(path, cloud_scheme=HCS_CLOUD_FILES_SCHEMA):
-        path_chunks = path.split(cloud_scheme + '://', 1)
-        if len(path_chunks) != 2:
-            raise RuntimeError('Unable to determine local path of [{}]'.format(path))
-        return '{}/{}'.format('/cloud-data', path_chunks[1])
-
-    @staticmethod
-    def _mkdir(path):
-        try:
-            os.makedirs(path)
-        except OSError as e:
-            if e.errno == errno.EEXIST and os.path.isdir(path):
-                pass
-            else:
-                return False
-        return True
-
-    @staticmethod
-    def get_selected_channels(params, channels):
-        selected_channels = []
-        selected_channel_ids = []
-        for chl in channels:
-            channel_params = params.get(chl['name'])
-            if channel_params is not None:
-                channel_params = channel_params.split(',')
-                if len(channel_params) != 5:
-                    raise RuntimeError("Channel should have 6 parameters")
-                channel = dict()
-                channel['id'] = chl['id']
-                channel['name'] = chl['name']
-                channel['r'] = int(channel_params[0])
-                channel['g'] = int(channel_params[1])
-                channel['b'] = int(channel_params[2])
-                channel['min'] = int(channel_params[3])
-                channel['max'] = int(channel_params[4])
-                selected_channel_ids.append(chl['id'])
-                selected_channels.append(channel)
-        if len(selected_channels) < 1:
-            raise RuntimeError("At least one channel is required")
-        return selected_channel_ids, selected_channels
-
-    @staticmethod
-    def get_ome_tiff_path(preview_dir, seq, by_field):
-        return preview_dir + '/{}/'.format(seq) + ('data.ome.tiff' if by_field else 'overview_data.ome.tiff')
-
-    @staticmethod
-    def get_channel_params(params, channel):
-        return next(filter(lambda p: p['channelId'] == channel, params))
-
-    @staticmethod
-    def merge_channels(images):
-        result = images[0]
-        for i in range(1, len(images)):
-            image = images[i]
-            r1, g1, b1 = image.split()
-            r2, g2, b2 = result.split()
-            r = np.minimum(255, np.add(np.array(r1), np.array(r2)))
-            g = np.minimum(255, np.add(np.array(g1), np.array(g2)))
-            b = np.minimum(255, np.add(np.array(b1), np.array(b2)))
-            result = Image.merge("RGB", (Image.fromarray(r), Image.fromarray(g), Image.fromarray(b)))
-        return result
-
-    @staticmethod
-    def change_intensity(intensity, contrast_min, contrast_max):
-        return (intensity - contrast_min) / max(0.0005, (contrast_max - contrast_min))
-
-    def color_image(self, image, channel):
-        image = self.change_intensity(image, channel['min'], channel['max'])
-        image = np.where(image <= 0, 0, 255 * image)
-        image = np.where(image <= 255, image, 255)
-        image = Image.fromarray(image)
-        grayscale_image = ImageOps.grayscale(image)
-        rgb = (channel['r'], channel['g'], channel['b'])
-        colored_image = ImageOps.colorize(grayscale_image, black='black', white=rgb)
-        return colored_image
-
     @staticmethod
     def _get_required_field(json_data, field_name):
         field_value = json_data.get(field_name)
         if field_value is None:
             raise RuntimeError("Field '%s' is required" % field_name)
         return field_value
+
+    @staticmethod
+    def get_required_field(json_data, field_name):
+        return HCSManager._get_required_field(json_data, field_name)
 
     @staticmethod
     def _collect_pipeline_tiff_outputs(pipeline_id, measurements_uuid):
